requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
blas=1.0=mkl
bokeh=1.1.0=py36_0
bzip2=1.0.6=h1de35cc_5
ca-certificates=2019.1.23=0
certifi=2019.3.9=py36_0
cffi=1.12.2=py36hb5b8e2f_1
click=7.0=py36_0
cloog=0.18.0=0
cloudpickle=0.8.1=py_0
cmake=3.14.0=haff7e42_0
cudatoolkit=9.0=h41a26b3_0
curl=7.64.1=ha441bb4_0
cycler=0.10.0=py36hfc81398_0
cython=0.29.7=py36h0a44026_0
cytoolz=0.9.0.1=py36h1de35cc_1
dask=1.2.0=py_0
dask-core=1.2.0=py_0
dask-jobqueue=0.1.1=py_0
distributed=1.27.0=py36_0
docrep=0.2.5=py_0
expat=2.2.6=h0a44026_0
fourier-accountant=0.12.0=pypi_0
freetype=2.9.1=hb4e5f40_0
gcc=4.8.5=8
gmp=6.1.2=hb37e062_1
hdf5=1.10.4=hfa1e0ec_0
heapdict=1.0.0=py36_2
icu=58.2=h4b95b61_1
intel-openmp=2019.3=199
isl=0.12.2=1
jbig=2.1=h4d881f8_0
jinja2=2.10.1=py36_0
jpeg=9b=he5867d9_2
kiwisolver=1.0.1=py36h0a44026_0
krb5=1.16.1=hddcf347_7
libcurl=7.64.1=h051b688_0
libcxx=4.0.1=hcfea43d_1
libcxxabi=4.0.1=hcfea43d_1
libedit=3.1.20181209=hb402a30_0
libffi=3.2.1=h475c297_4
libgfortran=3.0.1=h93005f0_2
libiconv=1.15=hdd342a3_7
libopenblas=0.3.3=hdc02c5d_3
libpng=1.6.36=ha441bb4_0
libssh2=1.8.2=ha12b0ac_0
libtiff=4.0.10=hcb84e12_2
libuv=1.26.0=h1de35cc_0
libxml2=2.9.9=hab757c2_0
libxslt=1.1.33=h33a18ac_0
locket=0.2.0=py36hca03003_1
lzo=2.10=h362108e_2
markupsafe=1.1.1=py36h1de35cc_0
matplotlib=3.0.3=py36h54f8f79_0
mkl=2019.3=199
mkl-include=2019.3=199
mkl_fft=1.0.10=py36h5e564d8_0
mkl_random=1.0.2=py36h27c97d8_0
mock=2.0.0=pypi_0
mpc=1.0.3=h7a72875_5
mpfr=3.1.5=h711e7fd_2
mpmath=1.1.0=py36_0
msgpack-python=0.6.1=py36h04f5b5a_1
ncurses=6.1=h0a44026_1
ninja=1.9.0=py36h04f5b5a_0
nodejs=10.13.0=h0a44026_0
numexpr=2.6.9=pypi_0
numpy=1.16.2=py36hacdab7b_0
numpy-base=1.16.2=py36h6575580_0
olefile=0.46=py36_0
openssl=1.1.1b=h1de35cc_1
packaging=19.0=py36_0
pandas=0.24.2=py36h0a44026_0
partd=0.3.10=py36_1
pbr=5.1.3=pypi_0
pillow=6.0.0=py36hb68e598_0
pip=19.0.3=py36_0
psutil=5.6.1=py36h1de35cc_0
pycparser=2.19=py36_0
pyparsing=2.4.0=py_0
python=3.6.8=haf84260_0
python-dateutil=2.8.0=py36_0
pytorch=1.0.1=py3.6_2
pytz=2019.1=py_0
pyyaml=5.1=py36h1de35cc_0
qt=5.9.7=h468cd18_1
readline=7.0=h1de35cc_5
rhash=1.3.8=ha12b0ac_0
scikit-learn=0.20.3=pypi_0
scipy=1.2.1=py36h1410ff5_0
setuptools=41.0.0=py36_0
six=1.12.0=py36_0
sklearn=0.0=pypi_0
sortedcontainers=2.1.0=py36_0
sqlite=3.27.2=ha441bb4_0
tables=3.5.1=pypi_0
tblib=1.3.2=py36hda67792_0
tk=8.6.8=ha441bb4_0
toolz=0.9.0=py36_0
torchvision=0.2.2=py_3
tornado=6.0.2=py36h1de35cc_0
unixodbc=2.3.7=h1de35cc_0
wget=1.20.1=h051b688_0
wheel=0.33.1=py36_0
xz=5.2.4=h1de35cc_4
yaml=0.1.7=hc338f04_2
zict=0.1.4=py36_0
zlib=1.2.11=h1de35cc_3
zstd=1.3.7=h5bba6e5_0
